refactor(Logging): 일부 클래스 패키지 이동
